Fix non-semantic version for git dependencies
If a git dependency is declared in an `idf_component.yml` with a version
not formatted as `major.minor.patch[-prerelease]` (where major, minor
and patch are integer numbers) a `ValueError: Invalid version string` is
returned by the semantic_version package.
It's very common for git projects to have tags starting with `v` or
release branches starting with `release-` and this error makes it
difficult to refer to a specific branch or tag.
Because the `semantic-version.Version` call is used only to evaluate if
the given version is semantic, the function should be surrounded by a
`try / except` and move on if the given version is not in the semantic
format.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>