feat: recognize cmakev2 projects cmake file
Currently, the add-dependency idf.py extension checks if a given
directory is a component directory. One of the checks involves
determining if the path is a project directory by testing if
CMakeLists.txt file includes `project.cmake`. However, cmakev2 projects
include a different `idf.cmake` file, so extend the existing
`is_component` check to consider `idf.cmake` as a project as well.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>